don't cast to float

+++ common/evaluators/bert_evaluator.py
@@ -72,7 +72,7 @@ class BertEvaluator(object):
                 if self.args.num_labels > 2:
                     logits = self.model(input_ids, input_mask, segment_ids)[0]
                 else:
-                    loss, logits = self.model(input_ids, input_mask, segment_ids, labels=label_ids.float())[:2]
+                    loss, logits = self.model(input_ids, input_mask, segment_ids, labels=label_ids)[:2]
 
             if self.args.is_multilabel:
                 predicted_labels.extend(F.sigmoid(logits).round().long().cpu().detach().numpy())
